# public/docs/api.yaml
openapi: 3.0.0
info:
  title: ARTube API
  description: ARTube 서버 REST API
  version: 0.1.0
  contact:
    email: "admin@isilab.kaist.ac.kr"
servers:
  - url: http://143.248.56.59:7000/api/v1
    description: API Test server
...
paths:
...
  /images:
    post:
      tags:
        - Image
      summary: 이미지 정보 업로드
      description: 앱을 통해 인식할 이미지 정보를 업로드합니다. 이 때 이미지 파일뿐만 아니라 이미지 파일 너비, 높이와 인쇄된 이미지(물리적 이미지)의 너비, 높이 정보를 입력해줘야 합니다. 물리적 이미지 크기는 단위는 m이며, 완벽히 정확한 크기를 입력하지 않아도 되지만 인쇄된 이미지의 너비, 높이 비율은 이미지 파일상의 너비, 높이 비율과 일치해야 합니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: 이미지 파일
                width:
                  type: integer
                  description: "이미지 파일 너비(단위: px)"
                height:
                  type: integer
                  description: "이미지 파일 높이(단위: px)"
                phyWidth:
                  type: number;
                  description: "물리적 이미지 너비(단위: m)"
                phyHeight:
                  type: number
                  description: "물리적 이미지 높이(단위: m)"
              required:
                - file
                - width
                - height
                - phyWidth
                - phyHeight
            encoding:
              file:
                contentType: image/png, image/jpeg, image/gif
            example:
              file: /media/dog.jpg
              width: 100
              height: 200
              phyWidth: 0.1
              phyHeight: 0.2
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/definitions/Image"
              example:
                _id: 5cdcec4b84cc2d4afb2cc3da
                userId: 5cda76af7a6a68609a521829
                filename: 5cda76af7a6a68609a5218491557982283354.jpg
                width: 200
                height: 300
                phyWidth: 0.2
                phyHeight: 0.3
                createdAt: "2019-05-14T08:05:03.602Z"
                updatedAt: "2019-05-14T08:05:03.602Z"
        "401":
          description: |
            1. HTTP 요청 헤더(Authorization)에 토큰이 없는 경우(Token is required)
            2. 토큰 유효 기간이 지난 경우(This token is expired)
            3. 유효한 토큰이 아닌 경우(This token is invalid)
    get:
      tags:
        - Image
      summary: 이미지 정보 리스트 가져오기
      description: 사용자가 업로드한 이미지 파일들 정보를 보여줍니다.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/definitions/Image"
        "401":
          description: |
            1. HTTP 요청 헤더(Authorization)에 토큰이 없는 경우(Token is required)
            2. 토큰 유효 기간이 지난 경우(This token is expired)
            3. 유효한 토큰이 아닌 경우(This token is invalid)
  /images/{id}:
    get:
      tags:
        - Image
      summary: 이미지 정보 가져오기
      description: 이미지 ID에 해당하는 이미지에 대한 정보를 보여줍니다.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: 이미지 ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/definitions/Image"
              example:
                _id: 3aeb28bd8b8g27409a791124
                userId: 5cda76af7a6a68609a521842
                filename: 5cda763aeb28bd8b8g27409a791124.jpg
                width: 100
                height: 200
                phyWidth: 0.1
                phyHeight: 0.2
                createdAt: "2019-05-14T08:05:03.602Z"
                updatedAt: "2019-05-14T08:05:03.602Z"
        "401":
          description: |
            1. HTTP 요청 헤더(Authorization)에 토큰이 없는 경우(Token is required)
            2. 토큰 유효 기간이 지난 경우(This token is expired)
            3. 유효한 토큰이 아닌 경우(This token is invalid)
    put:
      tags:
        - Image
      summary: 이미지 정보 업데이트
      description: 이미지 ID에 해당하는 이미지에 대한 정보를 업데이트합니다.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: 이미지 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                file:
                  type: string
                  description: 이미지 파일
                width:
                  type: integer
                  description: "이미지 파일 너비(단위: px)"
                height:
                  type: integer
                  description: "이미지 파일 높이(단위: px)"
                phyWidth:
                  type: number
                  description: "물리적 이미지 너비(단위: m)"
                phyHeight:
                  type: number
                  description: "물리적 이미지 높이(단위: m)"
            encoding:
              file:
                contentType: image/png, image/jpeg, image/gif
            example:
              file: /media/cat.jpg
              width: 120
              height: 240
              phyWidth: 0.24
              phyHeight: 0.48
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/definitions/Image"
              example:
                _id: 3aeb28bd8b8g27409a791124
                userId: 5cda76af7a6a68609a521842
                filename: 5cda763aeb28bd8b8g27409a791124.jpg
                width: 120
                height: 240
                phyWidth: 0.24
                phyHeight: 0.48
                createdAt: "2019-05-14T08:05:03.602Z"
                updatedAt: "2019-05-19T08:12:42.112Z"
        "401":
          description: |
            1. HTTP 요청 헤더(Authorization)에 토큰이 없는 경우(Token is required)
            2. 토큰 유효 기간이 지난 경우(This token is expired)
            3. 유효한 토큰이 아닌 경우(This token is invalid)
    delete:
      tags:
        - Image
      summary: 이미지 정보 삭제
      description: 이미지 ID에 해당하는 이미지 정보(이미지 파일 포함)를 삭제합니다. 삭제된 이미지 정보 및 파일은 복구할 수 없습니다.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: 이미지 ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
        "401":
          description: |
            1. HTTP 요청 헤더(Authorization)에 토큰이 없는 경우(Token is required)
            2. 토큰 유효 기간이 지난 경우(This token is expired)
            3. 유효한 토큰이 아닌 경우(This token is invalid)
  /videos:
    post:
      tags:
        - Video
      summary: 비디오 정보 업로드
      description: 이미지가 인식되었을 때 재생할 비디오 파일을 업로드합니다. 이 때 비디오 파일뿐만 아니라 비디오 파일 해상도 정보도 같이 업로드합니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: 비디오 파일
                width:
                  type: integer
                  description: "비디오 파일 너비(단위: px)"
                height:
                  type: integer
                  description: "비디오 파일 높이(단위: px)"
              required:
                - file
                - width
                - height
            encoding:
              file:
                contentType: video/mp4
            example:
              file: /media/dog.mp4
              width: 100
              height: 200
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/definitions/Video"
              example:
                _id: 5cdcec4b84cc2d4afb2cc3da
                userId: 5cda76af7a6a68609a521829
                filename: 5cda76af7a6a68609a5218491557982283354.mp4
                width: 800
                height: 600
                createdAt: "2019-05-14T08:05:03.602Z"
                updatedAt: "2019-05-14T08:05:03.602Z"
        "401":
          description: |
            1. HTTP 요청 헤더(Authorization)에 토큰이 없는 경우(Token is required)
            2. 토큰 유효 기간이 지난 경우(This token is expired)
            3. 유효한 토큰이 아닌 경우(This token is invalid)
    get:
      tags:
        - Video
      summary: 비디오 정보 리스트 가져오기
      description: 사용자가 업로드한 비디오 파일들 정보를 보여줍니다.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/definitions/Video"
        "401":
          description: |
            1. HTTP 요청 헤더(Authorization)에 토큰이 없는 경우(Token is required)
            2. 토큰 유효 기간이 지난 경우(This token is expired)
            3. 유효한 토큰이 아닌 경우(This token is invalid)
...
  /links:
    post:
      tags:
        - Link
      summary: 링크 정보 저장
      description: 앱에서 이미지가 인식되었을 때 비디오를 재생하기 위해서는 이미지와 비디오 간의 링크 정보가 필요합니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                  description: 이미지 ID
                video:
                  type: string
                  description: 비디오 ID
              required:
                - image
                - video
            example:
              image: ab78bc7d8s0099efsena8127
              video: c65a987b987a887a09a76bcc
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/definitions/Link"
              # example:
              #   _id: 5cdcec4b84cc2d4afb2cc3da
              #   userId: 2cda76af7a6a68609a521829
              #   filename:
              #   width: 800
              #   height: 600
              #   createdAt: '2019-05-14T08:05:03.602Z'
              #   updatedAt: '2019-05-14T08:05:03.602Z'
        "401":
          description: |
            1. HTTP 요청 헤더(Authorization)에 토큰이 없는 경우(Token is required)
            2. 토큰 유효 기간이 지난 경우(This token is expired)
            3. 유효한 토큰이 아닌 경우(This token is invalid)
    get:
      tags:
        - Link
      summary: 링크 정보 리스트 가져오기
      description: 사용자가 저장한 링크 정보를 보여줍니다.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/definitions/Link"
        "401":
          description: |
            1. HTTP 요청 헤더(Authorization)에 토큰이 없는 경우(Token is required)
            2. 토큰 유효 기간이 지난 경우(This token is expired)
            3. 유효한 토큰이 아닌 경우(This token is invalid)
  /links/{id}:
    get:
      tags:
        - Link
      summary: 링크 정보 가져오기
      description: 링크 ID에 해당하는 링크에 대한 정보를 보여줍니다.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: 링크 ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/definitions/Link"
              # example:
              #   _id: 3aeb28bd8b8g27409a791124
              #   userId: 5cda76af7a6a68609a521842
              #   filename: 5cda763aeb28bd8b8g27409a791124.mp4
              #   width: 100
              #   height: 200
              #   createdAt: '2019-05-14T08:05:03.602Z'
              #   updatedAt: '2019-05-14T08:05:03.602Z'
        "401":
          description: |
            1. HTTP 요청 헤더(Authorization)에 토큰이 없는 경우(Token is required)
            2. 토큰 유효 기간이 지난 경우(This token is expired)
            3. 유효한 토큰이 아닌 경우(This token is invalid)
    put:
      tags:
        - Link
      summary: 링크 정보 업데이트
      description: 링크 ID에 해당하는 링크에 대한 정보를 업데이트합니다.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: 링크 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                  description: 이미지 ID
                video:
                  type: string
                  description: 비디오 ID
            # example:
            #   file: /media/cat.mp4
            #   width: 800
            #   height: 600
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/definitions/Link"
              # example:
              #   _id: 3aeb28bd8b8g27409a791124
              #   userId: 5cda76af7a6a68609a521842
              #   filename: 5cda763aeb28bd8b8g27409a791124.mp4
              #   width: 800
              #   height: 600
              #   createdAt: '2019-05-14T08:05:03.602Z'
              #   updatedAt: '2019-05-19T08:12:42.112Z'
        "401":
          description: |
            1. HTTP 요청 헤더(Authorization)에 토큰이 없는 경우(Token is required)
            2. 토큰 유효 기간이 지난 경우(This token is expired)
            3. 유효한 토큰이 아닌 경우(This token is invalid)
    delete:
      tags:
        - Link
      summary: 링크 정보 삭제
      description: 링크 정보를 삭제합니다. 링크 정보를 삭제해도 링크에 포함된 이미지와 비디오 정보는 삭제되지 않습니다.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: 링크 ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
        "401":
          description: |
            1. HTTP 요청 헤더(Authorization)에 토큰이 없는 경우(Token is required)
            2. 토큰 유효 기간이 지난 경우(This token is expired)
            3. 유효한 토큰이 아닌 경우(This token is invalid)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
definitions:
...
  Image:
    properties:
      _id:
        type: string
        description: 이미지 ID
      userId:
        type: string
        description: 사용자 ID
      filename:
        type: string
        description: 이미지 파일명
      width:
        type: integer
        description: "이미지 파일 너비(단위: px)"
      height:
        type: integer
        description: "이미지 파일 높이(단위: px)"
      phyWidth:
        type: number
        description: "물리적 이미지 너비(단위: m)"
      phyHeight:
        type: number
        description: "물리적 이미지 높이(단위: m)"
      createdAt:
        type: string
        description: "생성 날짜/시간"
      updatedAt:
        type: string
        description: "마지막 수정 날짜/시간"
  Video:
    properties:
      _id:
        type: string
        description: 비디오 ID
      userId:
        type: string
        description: 사용자 ID
      filename:
        type: string
        description: 비디오 파일명
      width:
        type: integer
        description: "비디오 파일 너비(단위: px)"
      height:
        type: integer
        description: "비디오 파일 높이(단위: px)"
      createdAt:
        type: string
        description: "생성 날짜/시간"
      updatedAt:
        type: string
        description: "마지막 수정 날짜/시간"
  Link:
    properties:
      _id:
        type: string
        description: 링크 ID
      userId:
        type: string
        description: 사용자 ID
      image:
        type: string
        description: 이미지 ID
      video:
        type: string
        description: 비디오 ID
      createdAt:
        type: string
        description: "생성 날짜/시간"
      updatedAt:
        type: string
        description: "마지막 수정 날짜/시간"